requirements.txt
imageio==2.22.*
PyYAML==6.*
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
imageio=2.22
matplotlib==3.6
